backend/pyproject.toml
[project]
name = "behavior-stream-api"
version = "0.1.0"
description = "API para gerenciamento de atividades para psicólogos"
authors = [
    {name = "Fábio Henrique",email = "psifabiohenrique@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "django-cors-headers (>=4.3.1,<5.0.0)",
]

[tool.poetry]
packages = [{include = "behavior_stream_api", from = "../"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
